requirements.txt
# MetalQMS Requirements
# ======================

# Core Django
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-filter>=23.0
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.1

# Database
psycopg2-binary>=2.9.0

# File Processing
qrcode>=7.4.0
Pillow>=10.0.0
pypdf>=3.17.0
PyMuPDF>=1.23.0

# BPMN Workflow engine
django-viewflow>=2.2.0
celery>=5.3.0
redis>=5.0.0

# Telegram Bot API
python-telegram-bot>=20.0

# PDF Processing
reportlab>=4.0.0

# Monitoring and Logging
prometheus-client==0.20.0
django-prometheus==2.3.1
structlog==24.1.0
colorlog==6.8.2

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.3.0

# HTTP client for testing
requests>=2.31.0
psutil>=5.9.8